feat: create multi factor interface & move login resposne to it's own interface file
