+++ src/composables/types/authorization.ts
@@ -1,5 +1,3 @@
-import { SuccessResponse } from "./response";
-
 export interface userlogin {
     username: string,
     token?: string,
@@ -12,14 +10,3 @@ export interface formLogin {
     remember?: boolean
 }
 
-export interface LoginSuccess {
-    require_otp: false;
-    token: string;
-}
-
-export interface LoginRequiresOTP {
-    require_otp: true;
-    user_email: string;
-}
-
-export type LoginResponse = SuccessResponse<LoginSuccess | LoginRequiresOTP>;
