+++ src/composables/types/index.ts
@@ -1,3 +1,5 @@
 export type methodRequest = 'GET' | 'POST' | 'PUT' | 'DELETE';
 export * from "./authorization"
 export * from "./response"
+export * from "./multi-factor-auth"
+export * from "./login"
