+++ src/composables/types/multi-factor-auth.ts
@@ -1,0 +1,8 @@
+import { SuccessResponse } from "./response"
+
+export interface GetMultiFactorSuccess {
+    qr_code: string,
+    recovery_codes: string[]
+}
+
+export type MultiFactorResponse = SuccessResponse<GetMultiFactorSuccess>;
